+++ apps/example/.rnstorybook/preview.tsx
@@ -1,0 +1,24 @@
+import type { Preview } from "@storybook/react";
+import { withBackgrounds } from "@storybook/addon-ondevice-backgrounds";
+
+const preview: Preview = {
+  decorators: [withBackgrounds],
+  parameters: {
+    backgrounds: {
+      default: "plain",
+      values: [
+        { name: "plain", value: "white" },
+        { name: "warm", value: "hotpink" },
+        { name: "cool", value: "deepskyblue" },
+      ],
+    },
+    controls: {
+      matchers: {
+        color: /(background|color)$/i,
+        date: /Date$/,
+      },
+    },
+  },
+};
+
+export default preview;
