+++ apps/example/.rnstorybook/storybook.requires.ts
@@ -3,6 +3,7 @@ import { start, updateView, View } from '@storybook/react-native';
 
 import "@storybook/addon-ondevice-controls/register";
 import "@storybook/addon-ondevice-actions/register";
+import "@storybook/addon-ondevice-backgrounds/register";
 
 const normalizedStories = [
   {
