--- apps/example/.storybook/preview.tsx
@@ -1,17 +1,0 @@
-import type { Preview } from "@storybook/react";
-import { backgroundDecorator } from "@/developer/decorators";
-
-const preview: Preview = {
-  parameters: {
-    controls: {
-      matchers: {
-        color: /(background|color)$/i,
-        date: /Date$/,
-      },
-    },
-  },
-};
-
-export default preview;
-
-export const decorators = [backgroundDecorator];
